requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.7
django-filter==23.3
djangorestframework-simplejwt==5.3.0
Pillow==10.0.1
python-decouple==3.8
celery==5.3.4
redis==5.0.1
gunicorn==21.2.0
whitenoise==6.6.0
pytest-django==4.5.2
pytest==7.4.3
factory-boy==3.3.0
coverage==7.3.2

# Microsoft Graph API Integration
msgraph-core==0.2.2
azure-identity==1.15.0
msal==1.25.0
requests==2.31.0